Refactor MainActivity
